requirements.txt
requests
kafka-python
python-dotenv
fyers-apiv3
websocket-client
setuptools
psycopg2-binary
fastapi
uvicorn